# server/openapi.yaml
openapi: 3.1.0
info:
  title: Outing Briefing API
  description: Commute briefing service combining weather, air quality, and multiple traffic sources.
  version: 1.0.0
servers:
  - url: http://localhost:8787/api/v1
    description: Development server
paths:
  /profile:
    post:
      summary: Save user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
      responses:
        '200':
          description: Profile persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Fetch user profile
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /briefing:
    get:
      summary: Retrieve integrated commute briefing
      parameters:
        - in: query
          name: user_id
          required: false
          description: Stored profile identifier
          schema:
            type: string
        - in: query
          name: from
          required: false
          description: Origin coordinates formatted as "lat,lon"
          schema:
            type: string
            pattern: '^-?\d+\.\d+,-?\d+\.\d+$'
            example: "37.55,126.98"
        - in: query
          name: to
          required: false
          description: Destination coordinates formatted as "lat,lon"
          schema:
            type: string
            pattern: '^-?\d+\.\d+,-?\d+\.\d+$'
            example: "37.40,127.10"
        - in: query
          name: time
          required: false
          description: Desired departure time (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Briefing composed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Briefing'
        '400':
          description: Missing coordinates and profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Referenced profile missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /weather:
    get:
      summary: Retrieve weather brief
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
        - in: query
          name: lon
          required: true
          schema:
            type: number
        - in: query
          name: time
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Weather brief available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherBrief'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /air:
    get:
      summary: Retrieve air quality brief
      parameters:
        - in: query
          name: lat
          required: false
          schema:
            type: number
        - in: query
          name: lon
          required: false
          schema:
            type: number
        - in: query
          name: district
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Air quality brief available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirBrief'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /traffic/car:
    get:
      summary: Retrieve TMAP car route brief
      parameters:
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - $ref: '#/components/parameters/TimeParam'
      responses:
        '200':
          description: Car route brief
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficBrief'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /traffic/transit:
    get:
      summary: Retrieve TMAP transit route brief
      parameters:
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - $ref: '#/components/parameters/TimeParam'
      responses:
        '200':
          description: Transit route brief
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficBrief'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /traffic/expressway:
    get:
      summary: Retrieve expressway tollgate travel brief
      parameters:
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
        - $ref: '#/components/parameters/TimeParam'
      responses:
        '200':
          description: Expressway brief and tollgate metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpresswayTrafficResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Service operational
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'
components:
  parameters:
    FromParam:
      in: query
      name: from
      required: true
      description: Origin coordinates formatted as "lat,lon"
      schema:
        type: string
        pattern: '^-?\d+\.\d+,-?\d+\.\d+$'
        example: "37.55,126.98"
    ToParam:
      in: query
      name: to
      required: true
      description: Destination coordinates formatted as "lat,lon"
      schema:
        type: string
        pattern: '^-?\d+\.\d+,-?\d+\.\d+$'
        example: "37.40,127.10"
    TimeParam:
      in: query
      name: time
      required: false
      description: Requested departure time (ISO 8601)
      schema:
        type: string
        format: date-time
  schemas:
    SourceStatus:
      type: string
      enum: [ok, missing_api_key, upstream_error, timeout, bad_response]
    Coordinates:
      type: object
      properties:
        lat:
          type: number
        lon:
          type: number
      required: [lat, lon]
    TrafficStep:
      type: object
      properties:
        type:
          type: string
          enum: [drive, metro, bus, walk, bike]
        name:
          type: string
        duration_min:
          type: number
    TrafficBrief:
      type: object
      properties:
        source:
          type: string
          enum: [tmap, expressway]
        source_status:
          $ref: '#/components/schemas/SourceStatus'
        updated_at:
          type: string
          format: date-time
        mode:
          type: string
          enum: [car, transit, bike, walk]
        eta_minutes:
          type: number
        distance_km:
          type: number
        fare_krw:
          type: number
        transfers:
          type: integer
        steps:
          type: array
          items:
            $ref: '#/components/schemas/TrafficStep'
        congestion_level:
          type: string
          enum: [LOW, MID, HIGH]
        notes:
          type: array
          items:
            type: string
      required: [source, source_status, updated_at, mode]
    WeatherHourly:
      type: object
      properties:
        time:
          type: string
          format: date-time
        temp_c:
          type: number
        pop:
          type: number
        condition:
          type: string
      required: [time, temp_c, pop, condition]
    WeatherBrief:
      type: object
      properties:
        source:
          type: string
          example: kma
        source_status:
          $ref: '#/components/schemas/SourceStatus'
        updated_at:
          type: string
          format: date-time
        temp_c:
          type: number
        feels_like_c:
          type: number
        condition:
          type: string
        pop:
          type: number
          description: Precipitation probability (0-1)
        hourly:
          type: array
          items:
            $ref: '#/components/schemas/WeatherHourly'
        tmin_c:
          type: number
        tmax_c:
          type: number
        notes:
          type: array
          items:
            type: string
      required: [source, source_status, updated_at]
    AirBrief:
      type: object
      properties:
        source:
          type: string
          example: airkorea
        source_status:
          $ref: '#/components/schemas/SourceStatus'
        updated_at:
          type: string
          format: date-time
        pm10:
          type: number
        pm25:
          type: number
        grade:
          type: string
          enum: [good, normal, bad, verybad]
        advice:
          type: string
        station_name:
          type: string
        district:
          type: string
        notes:
          type: array
          items:
            type: string
      required: [source, source_status, updated_at]
    Recommendation:
      type: object
      properties:
        headline:
          type: string
        details:
          type: array
          items:
            type: string
        suggested_mode:
          type: string
          enum: [car, transit, bike, walk]
        leave_earlier_min:
          type: number
      required: [headline, details]
    Briefing:
      type: object
      properties:
        summary:
          type: string
        notices:
          type: array
          items:
            type: string
        weather:
          $ref: '#/components/schemas/WeatherBrief'
        air:
          $ref: '#/components/schemas/AirBrief'
        traffic:
          type: object
          properties:
            car:
              $ref: '#/components/schemas/TrafficBrief'
            transit:
              $ref: '#/components/schemas/TrafficBrief'
            expressway:
              $ref: '#/components/schemas/TrafficBrief'
          required: []
        recommendation:
          $ref: '#/components/schemas/Recommendation'
      required: [summary, notices, traffic]
    ExpresswayTrafficResponse:
      type: object
      properties:
        expressway:
          $ref: '#/components/schemas/TrafficBrief'
        meta:
          type: object
          properties:
            fromToll:
              type: string
            toToll:
              type: string
      required: [expressway, meta]
    ProfileInput:
      type: object
      properties:
        user_id:
          type: string
        preferred_mode:
          type: string
          enum: [car, transit, bike, walk]
        tz:
          type: string
        home:
          $ref: '#/components/schemas/ProfileLocation'
        work:
          $ref: '#/components/schemas/ProfileWorkLocation'
      required: [user_id, preferred_mode, tz, home, work]
    ProfileLocation:
      allOf:
        - $ref: '#/components/schemas/Coordinates'
        - type: object
          properties:
            label:
              type: string
            district:
              type: string
    ProfileWorkLocation:
      allOf:
        - $ref: '#/components/schemas/Coordinates'
        - type: object
          properties:
            label:
              type: string
    Profile:
      allOf:
        - $ref: '#/components/schemas/ProfileInput'
        - type: object
          properties:
            last_updated:
              type: string
              format: date-time
          required: [last_updated]
    HealthCheck:
      type: object
      properties:
        ok:
          type: boolean
        time:
          type: string
          format: date-time
        uptime:
          type: number
        version:
          type: string
      required: [ok, time]
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]